% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efs_operations.R
\name{efs_describe_mount_targets}
\alias{efs_describe_mount_targets}
\title{Returns the descriptions of all the current mount targets, or a specific
mount target, for a file system}
\usage{
efs_describe_mount_targets(MaxItems, Marker, FileSystemId,
  MountTargetId)
}
\arguments{
\item{MaxItems}{(Optional) Maximum number of mount targets to return in the response.
Currently, this number is automatically set to 10.}

\item{Marker}{(Optional) Opaque pagination token returned from a previous
\code{DescribeMountTargets} operation (String). If present, it specifies to
continue the list from where the previous returning call left off.}

\item{FileSystemId}{(Optional) ID of the file system whose mount targets you want to list
(String). It must be included in your request if \code{MountTargetId} is not
included.}

\item{MountTargetId}{(Optional) ID of the mount target that you want to have described
(String). It must be included in your request if \code{FileSystemId} is not
included.}
}
\description{
Returns the descriptions of all the current mount targets, or a specific
mount target, for a file system. When requesting all of the current
mount targets, the order of mount targets returned in the response is
unspecified.
}
\details{
This operation requires permissions for the
\code{elasticfilesystem:DescribeMountTargets} action, on either the file
system ID that you specify in \code{FileSystemId}, or on the file system of
the mount target that you specify in \code{MountTargetId}.
}
\section{Request syntax}{
\preformatted{efs$describe_mount_targets(
  MaxItems = 123,
  Marker = "string",
  FileSystemId = "string",
  MountTargetId = "string"
)
}
}

\examples{
# This operation describes all of a file system's mount targets.
\donttest{efs$describe_mount_targets(
  FileSystemId = "fs-01234567"
)}

}
\keyword{internal}
